feat(writers-room): vertical storyboard companion + UX cleanup
Replace the AI/Outline/Versions tabbed sidebar with an always-on
StoryboardPanel that renders the latest Adapt analysis as a vertical
column of scene cards, one per scene, each with its own image, summary,
characters, and per-scene Debug menu (Why this image / Check
characters / Editorial pass / Jump to prose). Stale banner appears when
the active draft hash differs from the analysis snapshot.

Click a scene card to jump the prose textarea to that scene. Falls back
to scene-summary substring or scene-index proportional scroll when the
LLM heading doesn't match the user's prose verbatim.

Header collapses to title + status + Save + Snapshot + a single (•••)
overflow menu organized into AI / Open / View sections. Outline,
Versions, Characters, and Analysis history each open as a slide-over
Drawer instead of stealing storyboard real estate.

Mobile: a Writing/Storyboard segmented toggle replaces the previous
stacked layout, fixing the double-scroll problem from c999def5.

AiPanel.jsx is decomposed into StoryboardPanel, SceneCard,
CharactersBible, AnalysisHistory, and sceneCardHelpers, each with a
single responsibility.
