+++ .changelog/NEXT.md
@@ -2,6 +2,8 @@
 
 ## Added
 
+- **Writers Room — vertical Storyboard companion + UX cleanup.** The AI/Outline/Versions tabbed sidebar is replaced with an always-on `StoryboardPanel` that shows the latest Adapt analysis as a vertical column of scene cards (heading, slugline, image, summary, character chips, visual prompt). Each card has a per-scene image regenerate button and a "Debug" overflow menu that scopes follow-up tools (Why this image / Check characters / Editorial pass / Jump to prose) to that one scene. Stale banner appears when the active draft's `contentHash` differs from the analysis's `sourceContentHash`, with a one-click re-run. Click a card heading to jump the prose textarea to that scene — falls back to scene-summary substring or scene-index proportional scroll when the LLM heading doesn't match the user's prose verbatim. The header collapses to title + status + Save + Snapshot + a single `(•••)` overflow menu (AI section: Run Adapt / Refresh characters / Editorial pass / Format pass · Open section: Outline / Versions / Characters [badge count] / Analysis history · View section: reading-theme toggle / Write-for-10 toggle). Outline, Versions, Characters, and Analysis history each open as a slide-over `Drawer` instead of competing with the storyboard for sidebar real estate. Mobile (<lg): a Writing/Storyboard segmented toggle replaces the previous stacked layout, fixing the double-scroll problem. The page-level Write-for-10 button moves into the work overflow menu so the Writers Room shell header is now title-only. New files: `StoryboardPanel.jsx`, `SceneCard.jsx`, `CharactersBible.jsx`, `AnalysisHistory.jsx`, `sceneCardHelpers.js`. Removed: `AiPanel.jsx` (decomposed into the above). Touched: `client/src/components/writers-room/WorkEditor.jsx`, `client/src/pages/WritersRoom.jsx`.
+
 - **Writers Room — character bible with image-gen prompt injection.** New `Characters` AI action extracts a per-work character roster (name, aliases, role, image-gen-ready physical description, personality, background) and merges it into an editable bible at `data/writers-room/works/<id>/characters.json`. Re-running the pass preserves any field the writer has already filled in (matched by canonical name OR alias, case-insensitive) and only fills blanks; prose-derived metadata (`firstAppearance`, `evidence`, `missingFromProse`) refreshes every run so the writer can see which fields the prose still doesn't support. Inline editor in the AI panel for create/edit/delete. The script-stage `SceneCard` now matches each scene's `characters` list against the bible and prepends `Featuring — <Name>: <physical description>` to the image-gen prompt before render, so the same character looks consistent across scenes. Matched character chips on the scene card show a port-accent border so the writer can see at a glance which characters in a scene have linked profiles. New routes: `GET/POST /api/writers-room/works/:id/characters`, `PATCH/DELETE /api/writers-room/works/:id/characters/:characterId`. New prompt stage `writers-room-characters` (sample + live). Touched: `server/services/writersRoom/{characters.js,evaluator.js}`, `server/routes/writersRoom.js`, `server/lib/{validation,writersRoomPresets}.js`, `client/src/components/writers-room/AiPanel.jsx`, `client/src/services/apiWritersRoom.js`, `data{,sample}/prompts/{stages/writers-room-characters.md,stage-config.json}`, `server/services/writersRoom/characters.test.js`.
 
 - **Collapsed-sidebar section flyout.** Hovering or focusing any multi-child section icon (Brain, Calendar, Chief of Staff, Comms, Create, Dev Tools, Digital Twin, MeatSpace, POST, Settings, System, Wiki) while the sidebar is collapsed opens a fixed-position popover to its right listing the section's children — so siblings (e.g. **Writers Room** under **Create**, which defaults to Media Gen) stay reachable without expanding the whole sidebar. Clicking a child navigates and closes the flyout; a 180 ms grace period lets the user move from icon → popover without it snapping shut. Renders via `position: fixed` so it escapes the nav scroller's `overflow-x-hidden` clipping. ARIA: `role="menu"` + `aria-haspopup="menu"` + `aria-expanded` on the trigger; keyboard focus on the icon also opens the flyout.
